moved things out into separate components to clean things up a bit
